# swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Gearvn Clone - OpenAPI 3.0
  description: |-
    Đây là một bản clone của Gearvn sử dụng OpenAPI 3.0 để mô tả API.

    Có một vài API mà chỉ admin mới có quyền truy cập (được chú thích ở từng API), vì vậy bạn cần đăng nhập với tài khoản admin để có thể truy cập vào những API đó.

    Các tác vụ như thêm thông báo, nhắn tin được thực hiện realtime với socket.io nên không có các api thực hiện những chức năng này. Nó được thực hiện thông qua giao tiếp giữa client và server với socket.io.
  termsOfService: http://swagger.io/terms/
  contact:
    email: haitrieu2524@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:4000
    description: Local server
...
paths:
...
  /users/verify-email:
    post:
      tags:
        - user
      summary: Xác thực email
      description: Xác thực email hay còn gọi là kích hoạt tài khoản để có thể sử dụng những chức năng của hệ thống cần xác thực người dùng như đổi mật khẩu, cập nhật thông tin cá nhân, ...
      operationId: verifyEmail
      security:
        - bearerAuth: []
      requestBody:
        description: Thông tin xác thực email
        content:
          application/json:
            schema:
              type: object
              properties:
                email_verify_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjUwODI5NTE2NGE4OTUwMmQ2MzJlZTdkIiwidmVyaWZ5IjowLCJyb2xlIjoxLCJ0b2tlbl90eXBlIjoyLCJpYXQiOjE2OTUwMzM2ODEsImV4cCI6MTY5NTYzODQ4MX0.fytsFO2uQrLjfvzalSY0la4q53DBmkpbHCcbhDEJfWg
        required: true
      responses:
        '200':
          description: Xác thực email thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Verify email succeed
                  data:
                    $ref: '#/components/schemas/SuccessAuthentication'
  /users/resend-email-verify:
    post:
      tags:
        - user
      summary: Gửi lại email xác thực
      description: Gửi lại email xác thực nếu người dùng chưa nhận được email xác thực hoặc email xác thực đã hết hạn
      operationId: resendEmailVerify
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Gửi lại email xác thực thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Resend email verify succeed
  /users/forgot-password:
    post:
      tags:
        - user
      summary: Quên mật khẩu
      description: Quên mật khẩu và gửi email reset password
      operationId: forgotPassword
      requestBody:
        description: Thông tin quên mật khẩu
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: haitrieu2527@gmail.com
        required: true
      responses:
        '200':
          description: Gửi email reset password thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Reset password request email has been sent, please check your email
...
  /users/reset-password:
    put:
      tags:
        - user
      summary: Đặt lại mật khẩu
      description: Đặt lại mật khẩu sau khi gửi yêu cầu quên mật khẩu và xác thực forgot_password_token thành công
      operationId: resetPassword
      requestBody:
        description: Thông tin đặt lại mật khẩu
        content:
          application/json:
            schema:
              type: object
              properties:
                forgot_password_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRlOWYzY2ZkNWNiZTU2OTIzNjdlNzA0IiwidmVyaWZ5IjoxLCJyb2xlIjoxLCJ0b2tlbl90eXBlIjozLCJpYXQiOjE2OTUwMzYyOTYsImV4cCI6MTY5NTY0MTA5Nn0.3Eq0x0AN5LIapgUDAAcAfaJyaL2FYQsKi6-lf_hX_Sk
                password:
                  type: string
                  example: 01225939705aA-
                confirm_password:
                  type: string
                  example: 01225939705aA-
        required: true
      responses:
        '200':
          description: Đặt lại mật khẩu thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Reset password succeed
                  data:
                    $ref: '#/components/schemas/SuccessAuthentication'
  /users/change-password:
    put:
      tags:
        - user
      summary: Đổi mật khẩu
      description: Thay đổi mật khẩu của tài khoản đang đăng nhập
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        description: Thông tin đổi mật khẩu
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                  example: 01225939704aA-
                password:
                  type: string
                  example: 01225939705aA-
                confirm_password:
                  type: string
                  example: 01225939705aA-
        required: true
      responses:
        '200':
          description: Đổi mật khẩu thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Change password succeed
  /users/refresh-token:
    post:
      tags:
        - user
      summary: Refresh token
      description: Refresh token để lấy access token mới khi access token đã hết hạn
      operationId: refreshToken
      requestBody:
        description: Thông tin refresh token
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRhNGM1N2U5ZTM2NzA3MjFhZTRiZTc3IiwidmVyaWZ5IjoxLCJyb2xlIjowLCJ0b2tlbl90eXBlIjoxLCJpYXQiOjE2OTUwMzMwODMsImV4cCI6MTY5NTAzMzY4M30.Ve9cBY7VRBYaDS9TqjW-bHkc6CY3tORsJN8NTr73l68
        required: true
      responses:
        '200':
          description: Refresh token thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Refresh token succeed
                  data:
                    $ref: '#/components/schemas/RefreshToken'
  /users/me:
    get:
      tags:
        - user
      summary: Lấy thông tin người dùng
      description: Lấy thông tin người dùng đang đăng nhập
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lấy thông tin người dùng thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Get info me succeed
                  data:
                    $ref: '#/components/schemas/User'
    patch:
      tags:
        - user
      summary: Cập nhật thông tin người dùng
      description: Cập nhật thông tin người dùng đang đăng nhập
      operationId: updateMe
      security:
        - bearerAuth: []
      requestBody:
        description: Thông tin cập nhật mới của người dùng nhập vào để cập nhật thông tin
        content:
          application/json:
            schema:
              type: object
              properties:
                fullname:
                  type: string
                  example: Trần Hải Triều
                avatar:
                  type: string
                  example: 977a9518243f1985fca916609.jpg
                gender:
                  type: integer
                  example: 0
                date_of_birth:
                  type: integer
                  example: 2000-11-17T17:00:00.000Z
                phone_number:
                  type: string
                  example: 0775939704
        required: true
      responses:
        '200':
          description: Cập nhật thông tin người dùng thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Update profile succeed
                  data:
                    $ref: '#/components/schemas/SuccessAuthentication'
  /users/viewed-product:
    post:
      tags:
        - user
      summary: Xem sản phẩm
      description: Thêm thông tin sản phẩm vào danh sách sản phẩm đã xem của người dùng
      operationId: addViewedProduct
      security:
        - bearerAuth: []
      requestBody:
        description: Thông tin sản phẩm
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: string
                  example: 60f9f3cfd5cbe5692367e704
        required: true
      responses:
        '200':
          description: Thêm sản phẩm vào danh sách sản phẩm đã xem thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Add or update viewed product succeed
    get:
      tags:
        - user
      summary: Lấy danh sách sản phẩm đã xem
      description: Lấy danh sách sản phẩm đã xem của người dùng đang đăng nhập
      operationId: getViewedProduct
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lấy danh sách sản phẩm đã xem thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Get viewed product succeed
                  data:
                    type: object
                    properties:
                      viewed_products:
                        type: array
                        items:
                          $ref: '#/components/schemas/ViewdProduct'
  /users:
    get:
      tags:
        - user
      summary: Lấy danh sách người dùng (chỉ có admin mới có quyền truy cập)
      description: Lấy danh sách tất cả người dùng trong hệ thống (chỉ có admin mới có quyền truy cập)
      operationId: getUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Trang hiện tại
          required: false
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          description: Số lượng người dùng trên một trang
          required: false
          schema:
            type: integer
            example: 10
        - name: gender
          in: query
          description: Giới tính
          required: false
          schema:
            type: integer
            example: 0
        - name: role
          in: query
          description: Vai trò
          required: false
          schema:
            type: integer
            example: 1
        - name: status
          in: query
          description: Trạng thái
          required: false
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description: Lấy danh sách người dùng thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Get users list succeed
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /users/customers:
    get:
      tags:
        - user
      summary: Lấy danh sách khách hàng (chỉ có admin mới có quyền truy cập)
      description: Lấy danh sách khách hàng trong hệ thống (chỉ có admin mới có quyền truy cập)
      operationId: getCustomers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Trang hiện tại
          required: false
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          description: Số lượng khách hàng trên một trang
          required: false
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Lấy danh sách khách hàng thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Get customers list succeed
                  data:
                    type: object
                    properties:
                      customers:
                        type: array
                        items:
                          $ref: '#/components/schemas/Customer'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  # Product
...
components:
  schemas:
    RefreshToken:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjUwODMwYzBlN2Q0ZTYyZDJmM2M5NjI2IiwidmVyaWZ5IjoxLCJyb2xlIjoxLCJ0b2tlbl90eXBlIjowLCJpYXQiOjE2OTUwMzY3NTQsImV4cCI6MTY5NTAzNzM1NH0.7cmEDRzcE0NuQ16iRX-XMqjWyFnU4ZzTYXh7Xc2glOI
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjUwODMwYzBlN2Q0ZTYyZDJmM2M5NjI2IiwidmVyaWZ5IjoxLCJyb2xlIjoxLCJ0b2tlbl90eXBlIjoxLCJleHAiOjE3MDM2NzY0NjEsImlhdCI6MTY5NTAzNjc1NH0.TRkCHfZ3zFxOxCFR5X0Xqq4D5_RjCHGxfFU1O5J7Ekg
    User:
      type: object
      properties:
        _id:
          type: string
          format: uuid
          example: 64a4c57e9e3670721ae4be77
        email:
          type: string
          example: haitrieu1811@gmail.com
        status:
          type: integer
          example: 0
        role:
          type: integer
          example: 0
        avatar:
          type: string
          example: 977a9518243f1985fca916609.jpg
        gender:
          type: integer
          example: 0
        verify:
          type: integer
          example: 0
        date_of_birth:
          type: integer
          example: 2000-11-17T17:00:00.000Z
        fullname:
          type: string
          example: Github Copilot
        phone_number:
          type: string
          example: 0775939704
        created_at:
          type: string
          format: ISO8601
          example: 2021-08-01T17:00:00.000Z
        updated_at:
          type: string
          format: ISO8601
          example: 2021-08-01T17:00:00.000Z
    SuccessAuthentication:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRhNGM1N2U5ZTM2NzA3MjFhZTRiZTc3IiwidmVyaWZ5IjoxLCJyb2xlIjowLCJ0b2tlbl90eXBlIjowLCJpYXQiOjE2OTUwMzMwODMsImV4cCI6MTY5NTAzMzY4M30.vmuKSMOMLUaFKMu0HQo4zHuS6lSSZKJZnQhfmP2YJfw
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRhNGM1N2U5ZTM2NzA3MjFhZTRiZTc3IiwidmVyaWZ5IjoxLCJyb2xlIjowLCJ0b2tlbl90eXBlIjoxLCJpYXQiOjE2OTUwMzMwODMsImV4cCI6MTcwMzY3MzA4M30.Ve9cBY7VRBYaDS9TqjW-bHkc6CY3tORsJN8NTr73l68
        user:
          $ref: '#/components/schemas/User'
    Pagination:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        page_size:
          type: integer
          example: 10
    ViewdProduct:
      type: object
      properties:
        _id:
          type: string
          example: 64c4addc50c11dd939ea30e9
        product:
          type: object
          properties:
            _id:
              type: string
              example: 60f9f3cfd5cbe5692367e704
            name_vi:
              type: string
              example: Chuột CoolerMaster MM711 RGB White Matte
            name_en:
              type: string
              example: CoolerMaster MM711 RGB White Matte Mouse
            thumbnail:
              type: string
              example: 60f9f3cfd5cbe5692367e704.jpg
            price:
              type: integer
              example: 1490000
            price_after_discount:
              type: integer
              example: 390000
            available_count:
              type: integer
              example: 10
            created_at:
              type: string
              format: ISO8601
              example: 2021-07-23T17:00:00.000Z
            updated_at:
              type: string
              format: ISO8601
              example: 2021-07-23T17:00:00.000Z
        created_at:
          type: string
          format: ISO8601
          example: 2021-08-01T17:00:00.000Z
        updated_at:
          type: string
          format: ISO8601
          example: 2021-08-01T17:00:00.000Z
...
    Customer:
      type: object
      properties:
        _id:
          type: string
          example: '64e9f3cfd5cbe5692367e704'
        email:
          type: string
          example: 'haitrieu2527@gmail.com'
        fullname:
          type: string
          example: 'Trần Hải Triều'
        phone_number:
          type: string
          example: '775939704'
        status:
          type: integer
          example: 0
        role:
          type: integer
          example: 1
        avatar:
          type: string
          example: '977a9518243f1985fca916609.jpg'
        gender:
          type: integer
          example: 0
        verify:
          type: integer
          example: 1
        orders_count:
          type: integer
          example: 0
        new_orders_count:
          type: integer
          example: 0
        processing_orders_count:
          type: integer
          example: 0
        delivering_orders_count:
          type: integer
          example: 0
        succeed_orders_count:
          type: integer
          example: 0
        canceled_orders_count:
          type: integer
          example: 0
        orders_total:
          type: integer
          example: 0
        new_orders_total:
          type: integer
          example: 0
        processing_orders_total:
          type: integer
          example: 0
        delivering_orders_total:
          type: integer
          example: 0
        succeed_orders_total:
          type: integer
          example: 0
        canceled_orders_total:
          type: integer
          example: 0
        addresses_count:
          type: integer
          example: 1
        date_of_birth:
          type: string
          format: date-time
          example: '2000-11-17T17:00:00.000Z'
        created_at:
          type: string
          format: date-time
          example: '2023-08-26T12:45:03.778Z'
        updated_at:
          type: string
          format: date-time
          example: '2023-09-19T01:21:57.490Z'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT